feat: enhance mockery command with git filter options
